Allow database handle to be provided from elsewhere
This functionality was documented to exist since inception, however
it has been broken since inception too: when the `init()` method
can't create a connection, an exception will be thrown and
initialization won't be finished.
This prohibits the inheriting class from calling its ancestor's
constructor to do the initialization other than the connection
instantiation.