Fix LazyAND condition failing when nested condition returns a string
As shown by the CheckReturn condition, conditions can return any value
that is interpreted as true or false, not just zero or one (or even a
number).

Since the result is used to check that more than a single condition was
evaluated, there's no need to use the condition result.