Fix instantiation of 'DBI handle from elsewhere' persisters
Requiring a DSN parameter in the workflow when the handle comes from
elsewhere, requires duplicate configuration. Instead, retrieve the
name of the driver from the handle. When no handle is available,
instantiate the ID generators from a provided driver name. If not
provided, fall back to instantiating the random generator, as is
already the case when the driver isn't one with built-in support.