Address issue #54 by making get_action() public in Workflow.pm
By renaming `Workflow->_get_action()` to `get_action()` and documenting
it to be public, consumers of Workflow can retrieve action object
instances and thereby the extra properties registered on them.

This commit introduces a new `_get_action()` method for backward
compatibility with 1.49 and earlier which simply forwards to its
public counterpart.