Fix bug found during review by @jonasbn
The Id generating classes referred to the `log` method, which neither
defined. This commit defines a `log` accessor for the attribute by
the same name.

Incidentally align the implementations of `SequenceId` and
`AutoGeneratedId` in that they both document their properties,
both define a `new` method and both define a `log` attribute now.