Re #68: Reduce the logging level of validation failures to 'debug'
As per @oliwel's request, change the logging of the validator's "error"
to DEBUG level: it's not really an error, it's a failure of the user to
provide compliant input (which is at *most* an informational message).

At the same time, add logging of condition errors as a TRACE level
message: these are not failures by any measure; they are merely
conditions evaluating to 'false'. When debugging, it might be handy
to be able to check these, but usually this is too chatty for debugging.
Putting the logging in TRACE instead.