Allow string interpolation while logging during initialization
During initialization, a truckload of actions are performed. It is my
expectation (to be verified) that a few string interpolations won't
make much difference in execution speed of initialization, even if the
resulting string ends up not being logged.

Two cases where logging might cause a more elaborate routine to create
the content to be logged (one call to Data::Dumper::Dumper() and the
other to 'join()' with a list of arguments) have been wrapped in a
sub{} (closure) to prevent execution when the output would not be logged.