Refactor 'autorun' handling into a loop (from recursion)
Very long workflows can cause deep stacks; OpenXPKI is experiencing
negative effects from this. However, in fact the situation is that
'autorun' can just as easily be handled with a loop.

This change factors most of 'execute_action' into the private function
'_execute_single_action' and incorporates '_auto_execute_state' into
'execute_action', which is now a loop which runs as long as the autorun
requirements are satisfied.