Fix #113: Support action attribute to be defined as part of State config
In the documentation for Workflow::Action, there's an explicit example
under the documentation of the C<new()> method which shows how to add
configuration (and index number) to an action class and copy that
configuration into an attribute of the Action's instance.

With this fix, that scenario is tested and has been fixed to work.