Fix #121: make Workflow::State->get_conditions() return all conditions (#122)
* Fix #121: make Workflow::State->get_conditions() return all conditions

Instead of returning a list of hashes and objects (which is undocumented
behaviour on a public function...), return a list of condition objects
as per the documentation.

While at it, clean up code that works around the undocumented behaviour.

* Make negated condition support transparent to the evaluator

Before this commit the condition evaluator special-cased
negated conditions. Instead of special-casing the conditions in
light of the evaluator, push negation support back to the
factory.