Document to override `init()` for processing of parameters (not `new()`) (#131)
The `new()` constructor is provided by the base class (`Workflow::Base`), which
calls `init()` to process the parameters provided. *Every* class in the library
should use `init()` for that purpose and many document this fact correctly.
Change `Workflow::Action`-s documentation to match the pattern.

Additionally, remove the **Note** about this being a hack; I'd call it a
design pattern, as the same is used for every other class in the system.