Eliminate undefined value warning
Before 1.55, this warning was hidden behind a `$log->is_info` check;
with that check removed, 1.55 now warns when no explicit `context_key`
is specified in the XML config.

Note that no context_key being specified, is the regular case when
the data field is a list of comma separated column names.