Change the contract between the persister and the workflow wrt history
The persister hard-codes the history class as does the workflow. This is in
high contrast with the rest of the library which allows user configuration
of nearly every aspect.

By changing the contract between the persister and its consumers (the workflow)
with respect to the way history is returned to the caller, flexibility is
created for the use of other history classes than Workflow::History (i.e.
derived classes).

Closes #124.