State internal accessors (#139)
* Add `_actions` accessor to State

Follow suit with other classes and other accessors in the State class

* Add `_conditions` accessor to State

Following the `_actions` accessor, add the `_conditions` accessor too

* Introduce `_next_state` accessor for next states for actions

Same as there is a `_conditions` accessor for actions' conditions,
we really want this accessor to separate this config out from
the action's true configuration.

* Remove State-consumed keys from state-action-config

State consumes the 'condition' and 'resulting_state' keys from the
Action configuration specified inside the State. Although these two
keys are specified *as part of* the Action config, they are in fact
State configuration relating to the Action, instead of true Action
configuration. This addresses an issue reported by @oliwell after the
1.55 release, where OpenXPKI was confused about unexpected parameters
being added to the action instantiation.

While at it, clean up the code around state-specific-action configuration
handling in general.

* Rename resulting_state related functions to next_state

... because the accessor is called `_next_state()`.

Co-authored-by: Jonas Brømsø <jonasbn@users.noreply.github.com>