Clean up documentation regarding tests (including test requirements) (#161)
* Check and update http(s) links from POD

This PR was inspired by a prior PR which removed a link to the "theoryx5" site,
because the link didn't resolve. With this PR, we have the machinery to verify
all links at once.

Michael Roberts changed his site layout; adapting the link.

* Clean up documentation regarding tests (including test requirements)

Note that this commit removes references to Perl::Critic testing in
Workflow's POD. There actually isn't such a file either in the repository
or in the generated dist. Hence removing the section referring to it.