Allow user-selectable history class (#168)
* Allow user-selectable history class

Closes #106. Moves the responsibility of instantiating history classes from the factory
and the persister into the Workflow class.

* Implement tests for custom workflow history configuration

Note that the actions were inserting history items of type 'Workflow::History'
which is the supertype of the configured history item.  Change the history
addition to simply provide the construction parameters of the history object
instead of providing the history object itself.

* Doc and string tweaks

- Merge two constant strings which would be concatenated later
- Indentation in a code example for better readability

* Doc change: Promote the use of history hashes for `add_history()`

Since the history class can be changed on the workflow, it's the
responsibility of Workflow to instantiate its history items. In our
examples, promote that pattern (instead of the older pattern where
the history class would be hard-coded).