Document that Workflow::Validator is an interface contract
Declaring the Workflow::Validator an interface contract as opposed to a
base class for validators allows *any* class to implement validations. This
addresses a long-standing desire of the OpenXPKI project to consolidate
the Condition and Validator concepts - in order to be able to share code
between the two - without actually consolidating them.