Enforce call to context->param as scalar
This is a follow up to the problem fixed by #195.
When a parameter given in the argument list of a validator does
not exist this is evaluated as list context and the empty list
does not change the runtime_args with the consequence that the
position of the arguments changes in the list, breaking the
functionality of the validator.