Remove internal parameter from public API
The `autorun` parameter was documented to be internal. It's also
no longer used due to migrating from recursion to a loop for the
`execute_action` function. Therefore, remove all references.

Also, Oliver Welter and I discussed the removal of the autorun
parameter to the observer calls: as "autorunning" isn't a criterion
to the state machine, it shouldn't be to an observer. So we decided
to drop it there too.