Enforce conditions to return a class object on evaluate
A strict type checking on the return value of the evaluate
call will prevent accidential decission errors due to bugs
in the condition classes.
The interface contract for conditions now enforces that an
instance of Workflow::Condition::IsTrue/::IsFalse is returned.
Any other value will cause the workflow to croak with an error